requirements.txt
absl-py==1.1.0
addict==2.4.0
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
black==19.10b0
bleach==5.0.1
cachetools==4.2.4
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
cycler==0.11.0
Cython==0.29.30
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.1
ftfy==6.1.1
google-auth==1.35.0
google-auth-oauthlib==0.4.6
graphviz==0.16
graspnetAPI==1.2.10
grpcio==1.47.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
imutils==0.5.3
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==4.7.2
jupyter-client==7.3.4
jupyter-core==4.11.1
jupyter-packaging==0.12.2
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.3.3
matplotlib-inline==0.1.3
memory-profiler==0.58.0
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mypy-extensions==0.4.3
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.19.4
oauthlib==3.2.0
object2urdf==0.12
open3d==0.15.2
open3d_plus==0.1.0
opencv-python==4.4.0.46
packaging==21.3
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.0.7
pycocotools==2.0.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.18.1
PyYAML==6.0
pyzmq==23.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
Send2Trash==1.8.0
Shapely==1.7.1
sniffio==1.2.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tensorboardX==2.5.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tomlkit==0.11.1
torch==1.7.1
torchvision==0.8.2
tornado==6.2
tqdm==4.57.0
traitlets==5.3.0
trimesh==3.12.9
urllib3==1.26.10
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
widgetsnbextension==3.6.1
zipp==3.8.1
